Update servers CMakeLists.txt
Due to how things are linked the google proto registry is linked twice, once for the server exe and once for the shared library, and a double register occurs.

This should be an easy fix to the OSX specific clang linking to prevent this, but for now just making the server library static is the easiest workaround.